added code to make a copy of the MRJob script with a unique module name
